flaggame/pyproject.toml
[tool.poetry]
name = "Flag Game Quiz"
version = "0.2.7"
description = "Flag Game Quiz built for Ohjelmistotekniikka K2023"
authors = ["joonarafael <joona.kettunen21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
invoke = "^2.0.0"
pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
coverage = "7.2.0"
pylint = "^2.17.2"
autopep8 = "^2.0.2"
2to3 = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
